Update connecton plugin to perserve the command build logic and provide support for Ansible 2.11 and lesser versions
